refactor(portal): move fetch-app-env fetcher + gql to scenes/common/app-env
