+++ web/scenes/common/app-env/graphql/server/fetch-app-env.generated.ts
